refactor(wp): move no-op tables code to deprecated file
