+++ db/DEPRECATEDwpdb.ts
@@ -192,7 +192,9 @@ export const DEPRECATEDgetPostType = async (
 }
 
 let cachedFeaturedImages: Map<number, string> | undefined
-export const getFeaturedImages = async (): Promise<Map<number, string>> => {
+export const DEPRECATEDgetFeaturedImages = async (): Promise<
+    Map<number, string>
+> => {
     if (cachedFeaturedImages) return cachedFeaturedImages
 
     const rows = await singleton.query(
@@ -206,4 +208,42 @@ export const getFeaturedImages = async (): Promise<Map<number, string>> => {
 
     cachedFeaturedImages = featuredImages
     return featuredImages
+}
+interface TablepressTable {
+    tableId: string
+    data: string[][]
+}
+export let cachedTables: Map<string, TablepressTable> | undefined
+export const DEPRECATEDgetTables = async (): Promise<
+    Map<string, TablepressTable>
+> => {
+    if (cachedTables) return cachedTables
+
+    const optRows = await singleton.query(`
+        SELECT option_value AS json FROM wp_options WHERE option_name='tablepress_tables'
+    `)
+
+    const tableToPostIds = JSON.parse(optRows[0].json).table_post
+
+    const rows = await singleton.query(`
+        SELECT ID, post_content FROM wp_posts WHERE post_type='tablepress_table'
+    `)
+
+    const tableContents = new Map<string, string>()
+    for (const row of rows) {
+        tableContents.set(row.ID, row.post_content)
+    }
+
+    cachedTables = new Map()
+    for (const tableId in tableToPostIds) {
+        const data = JSON.parse(
+            tableContents.get(tableToPostIds[tableId]) || "[]"
+        )
+        cachedTables.set(tableId, {
+            tableId: tableId,
+            data: data,
+        })
+    }
+
+    return cachedTables
 }
