+++ db/wpdb.ts
@@ -712,45 +712,6 @@ export const getTopics = async (cursor: string = ""): Promise<Topic[]> => {
     }
 }
 
-interface TablepressTable {
-    tableId: string
-    data: string[][]
-}
-
-let cachedTables: Map<string, TablepressTable> | undefined
-export const getTables = async (): Promise<Map<string, TablepressTable>> => {
-    if (cachedTables) return cachedTables
-
-    const optRows = await singleton.query(`
-        SELECT option_value AS json FROM wp_options WHERE option_name='tablepress_tables'
-    `)
-
-    const tableToPostIds = JSON.parse(optRows[0].json).table_post
-
-    const rows = await singleton.query(`
-        SELECT ID, post_content FROM wp_posts WHERE post_type='tablepress_table'
-    `)
-
-    const tableContents = new Map<string, string>()
-    for (const row of rows) {
-        tableContents.set(row.ID, row.post_content)
-    }
-
-    cachedTables = new Map()
-    for (const tableId in tableToPostIds) {
-        const data = JSON.parse(
-            tableContents.get(tableToPostIds[tableId]) || "[]"
-        )
-        cachedTables.set(tableId, {
-            tableId: tableId,
-            data: data,
-        })
-    }
-
-    return cachedTables
-}
-
 export const flushCache = (): void => {
     getBlogIndex.cache.clear?.()
-    cachedTables = undefined
 }
